refactor: rename feature folders to kebabcase
