+++ packages/api-reference/src/components/ApiReferenceLayout.vue
@@ -43,9 +43,9 @@ import ClassicHeader from '@/components/ClassicHeader.vue'
 import { Content } from '@/components/Content'
 import GettingStarted from '@/components/GettingStarted.vue'
 import MobileHeader from '@/components/MobileHeader.vue'
-import { ApiClientModal } from '@/features/ApiClientModal'
-import { useDocumentSource } from '@/features/DocumentSource'
-import { OPENAPI_VERSION_SYMBOL } from '@/features/DownloadLink'
+import { ApiClientModal } from '@/features/api-client-modal'
+import { useDocumentSource } from '@/features/document-source'
+import { OPENAPI_VERSION_SYMBOL } from '@/features/download-link'
 import { SearchButton } from '@/features/Search'
 import { Sidebar, useSidebar } from '@/features/sidebar'
 import { parse } from '@/helpers/parse'
