+++ packages/api-reference/src/components/Content/Tag/Tag.vue
@@ -14,7 +14,7 @@ import {
   SectionHeader,
   SectionHeaderTag,
 } from '@/components/Section'
-import { SpecificationExtension } from '@/features/SpecificationExtension'
+import { SpecificationExtension } from '@/features/specification-extension'
 import { useConfig } from '@/hooks/useConfig'
 import { useNavState } from '@/hooks/useNavState'
