+++ packages/api-reference/src/features/Operation/layouts/ClassicLayout.vue
@@ -18,10 +18,10 @@ import { Badge } from '@/components/Badge'
 import { HttpMethod } from '@/components/HttpMethod'
 import OperationPath from '@/components/OperationPath.vue'
 import { SectionAccordion } from '@/components/Section'
-import { ExampleRequest } from '@/features/ExampleRequest'
-import { ExampleResponses } from '@/features/ExampleResponses'
+import { ExampleRequest } from '@/features/example-request'
+import { ExampleResponses } from '@/features/example-responses'
 import type { Schemas } from '@/features/Operation/types/schemas'
-import { TestRequestButton } from '@/features/TestRequestButton'
+import { TestRequestButton } from '@/features/test-request-button'
 import { useConfig } from '@/hooks/useConfig'
 import {
   getOperationStability,
