+++ packages/api-reference/src/features/download-link/openapi-version-symbol.ts
