+++ packages/api-reference/src/features/example-responses/ExampleResponse.vue
