+++ packages/api-reference/src/features/specification-extension/SpecificationExtension.vue
